fix(repository): correct primary key type from Integer to Long in CreditRequirementRepository

+++ src/main/java/com/hyewon/wiseowl_backend/domain/requirement/repository/CreditRequirementRepository.java
@@ -3,5 +3,5 @@ package com.hyewon.wiseowl_backend.domain.requirement.repository;
 import com.hyewon.wiseowl_backend.domain.requirement.entity.CreditRequirement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CreditRequirementRepository extends JpaRepository<CreditRequirement, Integer>, CreditRequirementQueryRepository {
+public interface CreditRequirementRepository extends JpaRepository<CreditRequirement, Long>, CreditRequirementQueryRepository {
 }
